requirements.txt
asyncio
telethon
python-dotenv
pymongo
